requirements.txt
tensorflow-cpu>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
fastapi>=0.100.0
uvicorn>=0.22.0
joblib>=1.2.0
requests>=2.28.0
tqdm>=4.65.0
